feat:vistas pending, failure y success. Modularizacion del searchparams COMPONENTE
