+++ frontend/src/app/failure/page.tsx
@@ -1,8 +1,13 @@
- const Failure = ()=>{
+'use client'
+import { Suspense } from "react"
+import {SearchParams} from "../../components/SearchParams"
+
+
+const Failure = ()=>{
     return (
-        <div>
-            <h1>Compra fallida</h1>
-        </div>
+        <Suspense fallback ={<div>Loading...</div>}>
+            <SearchParams title="Compra fallida" description="Hubo un error con tu compra" />
+        </Suspense>
     )
 }
 export default Failure
